fix(session): bypass proxy-breaking probe connection
This branch is currently using a connection to probe for whether the
server uses http/1.1 or h2 and chooses a transport accordingly.

The problem is that this extra connection is breaking the proxy because
it uses net.Dial directly, so OONI doesn't work where a proxy is required.

As a fix, stop using the whole of netx inside session but rather let
us construct a netxlite transport.

I was already planning on doing that and this issue here is a good
excuse to see whether doing that makes sense.