doc: explain how to write ooni nettests
This is a preliminary commit and requires more work. I am going
to push this commit so others soon can start to review it.

This diff also contains a library of building blocks that is
instrumental to complete websteps. In fact, the tutorial I have
written here will eventually guide people to write their own
implementation of websteps.

I have not written any test for the new code so far. It is
also a stretch to call this new code. It is existing code that
has been written by me or @kelmenhorst that I have adapted
for the purpose of writing the tutorial.

The nice exercise of writing a tutorial is that you are more
easily frustrated if the code is difficult to use XD.

See https://github.com/ooni/ooni.org/issues/361.

See https://github.com/ooni/probe/issues/1733.