refactor: move KeyValueStore to internal/model
The time has now come to rationalize and review the code
inside of internal/engine/model.

The plan is basically to move around each interface or
data structure and review its usage.

In this case, I have removed redundant equivalent
declarations such that we only have one source of truth.