refactor: move i/netxlite/mocks to i/model/mocks
There are probably a couple of mocks that should not have been
moved as part of this change, we'll figure that out later.

BTW, this diff is part of https://github.com/ooni/probe/issues/1885
and https://github.com/ooni/probe-cli/pull/642. The latter PR
gives an high-level overview of the rationale of the changes I've
been implementing in this refactoring.