refactor: create common layer for archival
This diff starts addressing the first point mentioned in
the https://github.com/ooni/probe/issues/1957 issue:

> - [ ] both `internal/measurex` and `internal/engine/netx` have
[...] an archival data format. We should unify the two data formats
and ensure that the `ooni/spec` repository is in sync with them;

Here's what I did to produce it:

1. copy and adapt existing code from measurex that was
producing measurex's internal data format;

2. copy and adapt existing code from netx that was
producing the archival data format.

Basically, this code is measurex as far as measuring is
concerned (mainly because measurex is better written) and
netx for producing the archival data format (mainly because
such code has been tested much more).

In copying and adapting measurex code, I removed everything
that is not required by netx. Hence, this code will not
immediately work with measurex. To make it work, we need
to readd what was removed (mainly oddity computations).

Here's what I'm planning on doing next to ensure we finish
the first point of the above-mentioned issue:

1. adapt unit tests for the code I have already created;

2. replace code in netx to use the saver defined in this new
package instead of the one in netx;

3. replace code in the experiments as well;

4. replace code in measurex to use this package and make
sure we re-introduce what's needed (e.g., oddities).

The whole work will take a bunch of extra PRs. Doing this in
a single PR would result in a quite messy diff otherwise.

After all this work is done, the result is that:

1. we have more straightforward code for saving events
because we stop using the netx/trace.Event data structure,
which has the problem that, for every event, we use the
same data structure, so code is more like Python than Go;

2. we have reduced the amount of code required to implement
measurex by roughly 33% and are going in the direction or,
roughly, a reasonable merge of the good things inside of
netx and the better things inside measurex.

While there, also avoid using floating point for declaring
a constant inside of DASH, since Go does not care: every
constant is a number and is properly casted when used.